requirements.txt
streamlit
python-docx
pdfkit
openai
google-generativeai
